+++ ZipBuilder/Sources/FirebaseManifest/FirebaseManifest.swift
@@ -58,43 +58,3 @@ public struct Manifest {
   public let version: String
   public let pods: [Pod]
 }
-
-/// Struct describing Firebase pods to release.
-public struct Pod {
-  public let name: String
-  public let isClosedSource: Bool
-  public let isBeta: Bool
-  public let isFirebase: Bool
-  public let allowWarnings: Bool // Allow validation warnings. Ideally these should all be false
-  public let podVersion: String? // Non-Firebase pods have their own version
-  public let releasing: Bool // Non-Firebase pods may not release
-
-  init(_ name: String,
-       isClosedSource: Bool = false,
-       isBeta: Bool = false,
-       isFirebase: Bool = true,
-       allowWarnings: Bool = false,
-       podVersion: String? = nil,
-       releasing: Bool = true) {
-    self.name = name
-    self.isClosedSource = isClosedSource
-    self.isBeta = isBeta
-    self.isFirebase = isFirebase
-    self.allowWarnings = allowWarnings
-    self.podVersion = podVersion
-    self.releasing = releasing
-  }
-
-  public func podspecName() -> String {
-    return isClosedSource ? "\(name).podspec.json" : "\(name).podspec"
-  }
-
-  /// Closed source pods do not validate on Xcode 12 until they support the ARM simulator slice.
-  public func skipImportValidation() -> String {
-    if isClosedSource || name == "Firebase" {
-      return "-skip-import-validation"
-    } else {
-      return ""
-    }
-  }
-}
